Complete parts one and two of assignment
For part one: I struggled to get the HEX color working at first for the odd class, but it turns out I forgot the # in front. Also added the font-family code.

For part two: I completed the ID element in part 2 by adding the HEX color for text and changing font size.